separate library and cli
- make root package a cargo workspace
- move code to start client/server tunnels into lib.rs
- move rest of main.rs into wstunnel-cli bin package
- move "Wstunnel" config struct into bin package
  - the fields here are only applicable to the CLI
