+++ wstunnel-cli/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "wstunnel-cli"
+version = "10.1.8"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.95"
+
+clap = { version = "4.5.23", features = ["derive", "env"] }
+
+fdlimit = "0.3.0"
+
+tokio = { version = "1.42.0", features = ["full"] }
+
+tracing = { version = "0.1.41", features = ["log"] }
+tracing-subscriber = { version = "0.3.19", features = ["env-filter", "fmt", "local-time"] }
+
+wstunnel = { path = ".." }
